client: extracting StatItem for reusability
